Fri, Apr 26, 2024, 12:19 PM +08:00

+++ L4/ps4/filter-less/helpers.c
@@ -26,34 +26,34 @@ void sepia(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int tmpB = 0.272 * image[i][j].rgbtRed + 0.534 * image[i][j].rgbtGreen + 0.131 * image[i][j].rgbtBlue;
+            float tmpB = 0.272 * image[i][j].rgbtRed + 0.534 * image[i][j].rgbtGreen + 0.131 * image[i][j].rgbtBlue;
             if (tmpB > 255)
             {
                 image[i][j].rgbtBlue = 255;
             }
             else
             {
-                image[i][j].rgbtBlue = tmpB;
+                image[i][j].rgbtBlue = (int)tmpB;
             }
 
-            int tmpG = 0.349 * image[i][j].rgbtRed + 0.686 * image[i][j].rgbtGreen + 0.168 * image[i][j].rgbtBlue;
+            float tmpG = 0.349 * image[i][j].rgbtRed + 0.686 * image[i][j].rgbtGreen + 0.168 * image[i][j].rgbtBlue;
             if (tmpG > 255)
             {
                 image[i][j].rgbtGreen = 255;
             }
             else
             {
-                image[i][j].rgbtGreen = tmpG;
+                image[i][j].rgbtGreen = (int)tmpG;
             }
 
-            int tmpR = 0.393 * image[i][j].rgbtRed + 0.769 * image[i][j].rgbtGreen + 0.189 * image[i][j].rgbtBlue;
+            float tmpR = 0.393 * image[i][j].rgbtRed + 0.769 * image[i][j].rgbtGreen + 0.189 * image[i][j].rgbtBlue;
             if (tmpR > 255)
             {
                 image[i][j].rgbtRed = 255;
             }
             else
             {
-                image[i][j].rgbtRed = tmpR;
+                image[i][j].rgbtRed = (int)tmpR;
             }
         }
     }
